MAINT generalize flow, fix severe bug
OpenMLFlow parameters and components attribute are now of type
OrderedDict, with keys being the name of the parameter/component
and value the either the default value of the hyperparameter or
the actual component. This makes creating OpenMLFlows easier.
They can still be nicely uploaded.

Also, there was a bug in the deserialization, which returned
always the model of the serialized flow.