Refactor out different test cases to separate tests
The previous solution had two test conditions (strict and not strict)
and several scikit-learn versions, because of two distinct changes
within scikit-learn (the removal of min_impurity_split in 1.0, and the
restructuring of public/private models in 0.24).
I refactored out the separate test cases to greatly simplify the
individual tests, and I added a test case for scikit-learn>=1.0,
which was previously not covered.