Remove dtype checking for prediction comparison
It looks like the predictions loaded from an arff file are read as
floats by the arff reader, which results in a different type
(float v int). Because "equality" of values is already checked,
I figured dtype is not as imported. That said, I am not sure why
there are so many redundant comparisons in the first place?
Anyway, the difference should be due to pandas inference behavior,
and if that is what we want to test, then we should make a small
isolated test case instead of integrating it into every prediction
unit test.