[MNT] extend CI to newer python versions, deprecate python versions 3.8, 3.9 after EOL, marking further failing tests as `xfail` (#1579)
The CI runs only on python versions 3.8 and 3.9 both of which have already reached end of life.

This PR updates the python versions, deprecating any logic that runs tests on python versions 3.8 and 3.9, or `scikit-learn` versions of that age.

#### Metadata
Reference Issue: #1544 

Depends on https://github.com/openml/openml-python/pull/1584 fofr a fix, which should be merged first.

#### Details 

* The test matrix is updated to python versions, 3.10-3.13.
* Further failing tests are skipped using `mark.xfail` with  `reason="failures_issue_1544" ` for all the remaining failed tests (after #1572) in issue: #1544